test(SPV-1314): update visual regression tests (#1084)
