+++ src/__tests__/visual/_page-objects/defaults.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_WAIT_FOR_TIMEOUT = 2000;
